Moved node ops into common pkg
